Decouple RubyInstaller runtime libs and build libs.
It is desireable to lock the rubyinstaller gem for build
per bundler Gemfile.lock. However this can affect runtime
behaviour, if the RubyInstaller version is different from
the rubyinstaller gem.

We therefore need to clearly separate between runtime and
build gem. Runtime files are not shipped as gem, but as plain
files in the $LOAD_PATH search path, so that they
are always consistent with the rest of the ruby installation.
They also use their own namespace RubyInstaller::Runtime.

In contrast build files are shipped per gem and are not bound to
a specific RubyInstaller runtime version. All build related
classes are under RubyInstaller::Build. Both namespaces can
coexist in different versions within the same ruby VM.

Some library files are used at build and runtime. These files
are stored in "ruby_installer/build" only (in git), but are rewritten
to "ruby_installer/runtime" in the sandbox task. They also use a
different namespace then (RubyInstaller::Runtime vs.
RubyInstaller::Build).

This also changes the repository file structure a lot.