Change the way how static files from git/gem are loaded.
Static files are now handled similar to an overlay filesystem.
They are preferably used from the current working directory
and if now found, are taken from the git working directory or
the rubyinstaller-build.gem depending on which is used by the
package.

This way each package can override arbitrary files of the
rubyinstaller-build.gem.

Any generated or copied files are stored in the current working
directory. This way each package has it's own set of temporary files.