Set Rubys Encoding.default_external to UTF-8.
Rubys default_external setting depends on the language/region of the local
Windows installation. It typically doesn't fit to the encoding of the
text files to read. However since UTF-8 is very common these days (and the
UTF-8 everywhere strategy), it's better to set UTF-8 as reasonable
default.

This also avoids a typical interop issue with other operating systems.