Don't use overlay mechanism for PKGBUILD dependency files.
These files were copied only once and were recognized as overloaded
for subsequent runs, so that changes on the original files were
not forwarded. Since the overlay mechanism doesn't make much sense
for these files, we handle them as pure gem files now.
This makes sure, that they are updated in the local directory.