Sign files and upload them directly from Appveyor to Github
This is triggered by pushing a git tag kind of "rubyinstaller-2.4.1-2" to github.
Appveyor will subsequently start a build of the tagged state.
When the binary artefacts are built, they get signed by the "ri2-ci" GPG key.
The key is stored encrypted in the git repository.
Then all artefacts are uploaded to github.

Since x64 and x86 builds are in separate Appveyor runs, only the first run
creates the github release and subsequent runs add artefacts only.