Uninstall old RubyInstaller version before install
Previous RubyInstaller versions simply overwrote files in the install
folder, but kept files, which are removed or renamed in the new version.

One example is, when a default gem is updated. In this case an old file
is kept, which could potentially lead to failues:
lib/ruby/gems/2.4.0/specifications/default/openssl-2.0.3.gemspec

This patch automatically triggers an uninstall of the previous
RubyInstaller version, just before the new installation starts.