Add "rake docker:builds" to reliable build installers
It makes use of docker containers, so that the build doesn't depend
on any local states. The only input is the git repository.
It uses a mapped local cache to avoid unnecessary downloads from the
internet. However integrity of these files is verified by gpg.

Intention of this change is to migrate the current Appveyor based build
to an on premises solution. This allows to digitally sign executables
within the installer like ruby.exe . The current workflow only signs
the installer file produced on Appveyor, but not the files inside.
Docker provides a comparable isolation to Appveyors virtual machines
and allows access to a smart card for signatures.