Fix addition of ruby_installer/runtime files from gem
gemspec.files can be used when rake is running in the rubyinstaller2 project tree.
However if rubyinstaller/build is used as a gem, then we can't easily load the gemspec from file, because it depends on "git
ls-files" but the gem doesn't ship the ".git" directory.
On the other hand gemspec.files returns [] for a loaded gem.
Therefore we use Dir.glob to retrieve the runtime file list.