Fix irb hook in ruby-3.1
Ruby-3.1 changed to irb.bat as binstub, so that the hook was no longer active.
It is generated in https://github.com/oneclick/rubyinstaller2/blob/master/recipes/sandbox/20-extend-irb.rake#L5
to enable automatic creation of the .irbrc file and encoding conversion of the .irb_history file.
In ruby31 there is no need any longer for the automatic creation of the .irbrc file, since the irb has suitable defaults these days.
The only task that the irb.cmd is still useful, is the encoding conversion to UTF-8 when upgrading from older ruby versions before mid-2020.
Since that isn't too far in the past, this commit re-enables the irb hook.

In some years the irb-patching can be removed entirely.

Also change the start menu entry to irb.bat in ruby-3.1.