Avoid UNICODE characters in TMP env var
Some people have Unicode characters in their TMP path.
This is not properly handled by gcc, when it creates temporary files, so that it fails to compile.

As a work-around we verify that TMP contains only ACSII characters, and if not set it to MSYS2's /tmp.
This variable es NOT restored by "ridk disable", because the cmd and powershell mechanism to set variables doesn't support Unicode either.

Fixes #320