Move SSL certs directory to <ruby>/bin/etc/ssl
This is due to a path relocation change in MSYS2:
  https://github.com/msys2/MINGW-packages/pull/16141

Relocation of the certs directory is now done based on the path of the libssl-3-x64.dll instead of the running executable ruby.exe .
Since we store the DLL in bin/ruby_builtin_dlls, the certs directory is expected to be under <ruby>/bin and not under <ruby> root directory.

There is no good way to adjust the relocated path for RubyInstaller other than forking the build script.
To avoid additional maintanace burden and since we want to keep using latest OpenSSL versions, RubyInstaller follows this change, even if it's certainly not the best place.

Fixes #337