Move the SSL CA certs directory from ruby/bin/etc/ssl to ruby/ssl
It was moved in ruby-3.2 to etc/ssl and short time later to bin/etc/ssl because of changes in MSYS2-MINGW.
Both directory names are suboptimal.
Now we have to move the libssl.dll, due to the SxS changes, so that the certs directory would change another time.
To avoid another move this is now solved with a link (junction) to the original directory ruby/ssl .

Downside is that mklink always creates a junction with absolute path, even when only a relative path is given to the command.
It also needs a NTFS drive.
When the ruby installation is moved to a USB stick, then the CA certs are no longer found.
It can be manually fixed by moving the certs directory ruby/ssl into ruby/lib/ruby/<vers>/etc .