Reverse the junction/link to "ssl" dir
Now the certs are stored under "<ruby>/lib/..." and "<ruby>/ssl" is linked to this directory per innosetup installer.

Before this commit, the link was the other way around, but junctions can not be stored/restored in a 7z archive.
This led to certs not found when 7z archive was used, like in ruby/setup-ruby.

Fixes #405