Remove installed gems and MSYS2 from devkit by the uninstaller
So far the uninstaller only removed the ruby install files, but kept the installed gems and MSYS2.
This is the right behaviour for an update installation, which uninstalls the previous ruby in the install directory first.
But it is not expected, when uninstalling from "Add/Remove Programs".
Also Microsoft Store requires to remove all related files.